Handle comments:
Fix errors for lint
Add comments in readme.
