+++ README.md
@@ -1,2 +1,9 @@
 # esper-api-sample-code
 Sample Codes demonstrating use of Esper API's
+
+Please edit the python file to update your Enterprise Related information
+under CONFIGURATION: endpoint-name, API-KEY and ENTERPRISE-ID
+
+CONFIGURATION.host = 'https://<endpoint-name>-api.shoonyacloud.com/api'
+CONFIGURATION.api_key['Authorization'] = '<API-KEY>'
+ENTERPRISE_ID = '<ENTERPRISE-ID>'
